+++ Fit/Utils/MinimizerUtils.h
@@ -12,19 +12,20 @@
 //
 // ************************************************************************** //
 
-#ifndef  MINIMIZERUTILS_H
-#define  MINIMIZERUTILS_H
+#ifndef MINIMIZERUTILS_H
+#define MINIMIZERUTILS_H
 
 #include "WinDllMacros.h"
+#include <map>
 #include <string>
 #include <vector>
-#include <map>
 
 //! Utility functions for fit library
 
-namespace MinimizerUtils {
+namespace MinimizerUtils
+{
 
-BA_CORE_API_ std::string toString(const std::vector<std::string>& v, const std::string& delim="");
+BA_CORE_API_ std::string toString(const std::vector<std::string>& v, const std::string& delim = "");
 
 BA_CORE_API_ std::map<int, std::string> gslErrorDescriptionMap();
 
@@ -32,7 +33,8 @@ BA_CORE_API_ std::string gslErrorDescription(int errorCode);
 
 BA_CORE_API_ bool numbersDiffer(double a, double b, double tol);
 
-BA_CORE_API_ std::string sectionString(const std::string& sectionName="", int report_width=80);
+BA_CORE_API_ std::string sectionString(const std::string& sectionName = "",
+                                       size_t report_width = 80);
 
 } // namespace MinimizerUtils
