+++ Fit/Utils/StringUtils.cpp
@@ -49,10 +49,10 @@ bool StringUtils::matchesPattern(const std::string& text, const std::string& wil
 }
 
 //! Returns string right-padded with blanks.
-std::string StringUtils::padRight(const std::string& name, int length)
+std::string StringUtils::padRight(const std::string& name, size_t length)
 {
     std::string result = name;
-    result.resize(length,' ');
+    result.resize(length, ' ');
     return result;
 }
 
@@ -64,10 +64,10 @@ std::vector<std::string> StringUtils::split(const std::string& text, const std::
     return tokens;
 }
 
-void StringUtils::replaceItemsFromString(
-    std::string& text, const std::vector<std::string>& items, const std::string& replacement)
+void StringUtils::replaceItemsFromString(std::string& text, const std::vector<std::string>& items,
+                                         const std::string& replacement)
 {
-    for(size_t i=0; i<items.size(); ++i)
+    for (size_t i = 0; i < items.size(); ++i)
         boost::replace_all(text, items[i], replacement);
 }
 
@@ -75,17 +75,19 @@ std::string StringUtils::join(const std::vector<std::string>& joinable, const st
 {
     std::string result;
     size_t n = joinable.size();
-    if (n==0) return result;
-    for(size_t i=0; i<n-1; ++i)
+    if (n == 0)
+        return result;
+    for (size_t i = 0; i < n - 1; ++i)
         result += joinable[i] + joint;
-    result += joinable[n-1];
+    result += joinable[n - 1];
     return result;
 }
 
-std::string StringUtils::removeSubstring(const std::string &text, const std::string &substr)
+std::string StringUtils::removeSubstring(const std::string& text, const std::string& substr)
 {
     std::string result = text;
-    for(std::string::size_type i=result.find(substr); i!=std::string::npos; i=result.find(substr))
+    for (std::string::size_type i = result.find(substr); i != std::string::npos;
+         i = result.find(substr))
         result.erase(i, substr.length());
     return result;
 }
