+++ Fit/Utils/StringUtils.h
@@ -23,19 +23,20 @@
 
 //! Utility functions to analyze or modify strings.
 
-namespace StringUtils {
+namespace StringUtils
+{
 
 //! Returns true if text matches pattern with wildcards '*' and '?'.
 BA_CORE_API_ bool matchesPattern(const std::string& text, const std::string& wildcardPattern);
 
-BA_CORE_API_ std::string padRight(const std::string& name, int length);
+BA_CORE_API_ std::string padRight(const std::string& name, size_t length);
 
 //! Split string into vector of string using delimeter.
 BA_CORE_API_ std::vector<std::string> split(const std::string& text, const std::string& delimeter);
 
 //! Replaces all occurences of items from string text with delimiter
-BA_CORE_API_ void replaceItemsFromString(
-    std::string& text, const std::vector<std::string>& items, const std::string& replacement="");
+BA_CORE_API_ void replaceItemsFromString(std::string& text, const std::vector<std::string>& items,
+                                         const std::string& replacement = "");
 
 //! Returns string obtain by joining vector elements
 BA_CORE_API_ std::string join(const std::vector<std::string>& joinable, const std::string& joint);
@@ -44,16 +45,12 @@ BA_CORE_API_ std::string join(const std::vector<std::string>& joinable, const st
 BA_CORE_API_ std::string removeSubstring(const std::string& text, const std::string& substr);
 
 //! Returns scientific string representing given value of any numeric type.
-template <typename T> std::string scientific(const T value, int n = 10); // implementation below
+template <typename T> std::string scientific(const T value, int n = 10);
 
 //! Returns new string which is lower case of text.
 BA_CORE_API_ std::string to_lower(std::string text);
 
 } // namespace StringUtils
-
-// ************************************************************************** //
-// implementation
-// ************************************************************************** //
 
 template <typename T> std::string StringUtils::scientific(const T value, int n)
 {
